+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_archive_utils.py
@@ -7,6 +7,7 @@ import tarfile
 import os
 import re
 import codecs
+from io import open
 import requests
 from knack.log import get_logger
 from knack.util import CLIError