+++ src/azure-cli-testsdk/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+0.2.2
++++++
+* Minor fixes
 
 0.2.1
 +++++