+++ src/command_modules/azure-cli-servicefabric/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+0.1.6
++++++
+* Minor fixes
 
 0.1.5
 +++++