+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+
+2.2.6
+++++++
+* `vm/vmss create`: enforce disk caching mode be `None` for Lv/Lv2 series of machines
+* `vm create`: update supported size list supporting networking accelerator
+* `disk update`: expose strong typed arguments for ultrassd iops and mbps configs
+
 2.2.5
 ++++++
 * Fix SDK issue that caused Homebrew instllation to fail.